+++ src/main/java/org/devkor/apu/saerok_server/domain/auth/api/dto/response/AccessTokenResponse.java
@@ -2,13 +2,13 @@ package org.devkor.apu.saerok_server.domain.auth.api.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-@Schema(description = "JWT 토큰 및 회원가입 상태 응답")
-public record LocalJwtResponse(
+@Schema(description = "액세스 토큰 및 회원가입 상태 응답")
+public record AccessTokenResponse(
 
         @Schema(description = "액세스 토큰", example = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...")
         String accessToken,
 
-        @Schema(description = "회원가입 진행 상태", example = "COMPLETED")
+        @Schema(description = "회원가입 상태", example = "COMPLETED")
         String signupStatus
 ) {
 }
